% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-function.R
\name{plot_something}
\alias{plot_something}
\title{Takes some fake data and modifies it. Demo purposes only.}
\usage{
plot_something(input)
}
\arguments{
\item{input}{Some data}
}
\description{
A demonstration function that takes a data frame (input) and modifies it to produce an (output).
}
